Changing archetype based on best practices
