+++ archetype/simple-service/src/main/resources/archetype-resources/service/src/main/java/service/__name__Module.java
@@ -5,31 +5,13 @@ package ${package}.service;
 
 import ${package}.service.resource.HelloWorldResourceImpl;
 import ${package}.api.HelloWorldResource;
-import com.cerner.beadledom.metadata.BuildInfo;
-import com.cerner.beadledom.metadata.ServiceMetadata;
-import com.cerner.beadledom.resteasy.ResteasyModule;
-import com.google.inject.AbstractModule;
-import com.google.inject.Provides;
-import com.wordnik.swagger.config.SwaggerConfig;
-import com.wordnik.swagger.core.SwaggerSpec;
+import com.google.inject.PrivateModule;
 
-public class ${name}Module extends AbstractModule {
+public class ${name}Module extends PrivateModule {
 
   protected void configure() {
-    install(new ResteasyModule());
-
-    BuildInfo buildInfo = BuildInfo.load(getClass().getResourceAsStream("build-info.properties"));
-    bind(BuildInfo.class).toInstance(buildInfo);
-    bind(ServiceMetadata.class).toInstance(ServiceMetadata.create(buildInfo));
-
     bind(HelloWorldResource.class).to(HelloWorldResourceImpl.class);
-  }
 
-  @Provides
-  SwaggerConfig provideSwaggerConfig(ServiceMetadata serviceMetadata) {
-    SwaggerConfig config = new SwaggerConfig();
-    config.setApiVersion(serviceMetadata.getBuildInfo().getVersion());
-    config.setSwaggerVersion(SwaggerSpec.version());
-    return config;
+    expose(HelloWorldResource.class);
   }
 }
